+++ CHANGELOG.md
@@ -1,5 +1,17 @@
 # Change Log
 
+## [0.2.1] - 2026-02-15
+
+### Changed
+
+* Updated all dependencies.
+
+* `slow_tests` feature for slow tests.
+
+### Improved
+
+* Implemented `DefaultEstimator::into_owned`.
+
 ## [0.2.0] - 2025-10-16
 
 ### Changed